+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # RealmJoin Remediation Scripts Changelog
 
+## 2025-11-13
+- Added Align device name with Autopilot profile
+
 ## 2025-10-27
 - Updated Remove HP applications